rename: constants.ts 파일 하나로 합치기
