+++ src/app/constants/constants.ts
@@ -1,2 +1,4 @@
 export const GIFTBAG_NAME_MAX_LENGTH = 20;
 export const GIFT_NAME_MAX_LENGTH = 20;
+
+export const GIFT_SELECT_REASON_MAX_LENGTH = 100;
